ddos whitelist functions
