+++ txpool/query.go
@@ -52,39 +52,3 @@ func (p *TxPool) GetTxs(inclQueued bool) (
 func (p *TxPool) Pending() map[types.Address][]*types.Transaction {
 	return p.accounts.poolPendings()
 }
-
-const (
-	DDosWhiteList = "whitelist"
-	DDosBlackList = "blacklist"
-)
-
-// GetDDosContractList shows current white list and black list contracts
-func (p *TxPool) GetDDosContractList() map[string]map[types.Address]int {
-	var (
-		ret    = make(map[string]map[types.Address]int, 2)
-		blacks = make(map[types.Address]int)
-		whites = make(map[types.Address]int)
-	)
-
-	p.ddosContracts.Range(func(key, value interface{}) bool {
-		addr, _ := key.(types.Address)
-		count, _ := value.(int)
-
-		if isCountExceedDDOSLimit(count) {
-			blacks[addr] = count
-		}
-
-		return true
-	})
-
-	p.ddosWhiteList.Range(func(key, value interface{}) bool {
-		addr, _ := key.(types.Address)
-		whites[addr] = 0
-		return true
-	})
-
-	ret[DDosBlackList] = blacks
-	ret[DDosWhiteList] = whites
-
-	return ret
-}
